# src/main/docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /empleado/modi/{id}:
    put:
      tags:
      - empleado-controller
      operationId: modificarEmpleado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Empleado'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
  /empleado:
    post:
      tags:
      - empleado-controller
      operationId: guardarEmpleado
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Empleado'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Empleado'
...
  /empleado/{id}:
    get:
      tags:
      - empleado-controller
      operationId: consultarEmpleado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Empleado'
...
  /empleado/borrar/{id}:
    delete:
      tags:
      - empleado-controller
      operationId: eliminarEmpleado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
...
    Empleado:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nombre:
          type: string
        direccion:
          type: string
        telefono:
          type: integer
          format: int32
        cargo:
          type: string